core.controller: fix move_frame method
git-svn-id: svn+ssh://svn.gna.org/svn/advene/trunk@4306 3fb6f576-06d4-0310-ad11-a4b695aa9f05

+++ lib/advene/core/controller.py
@@ -1749,7 +1749,7 @@ class AdveneController:
         elif p.status != p.PauseStatus:
             self.update_status('start')
             self.update_status('pause')
-        self.move_position (1000 / 25 * number_of_frames, notify=False)
+        self.move_position (int(1000 / 25.0 * number_of_frames), notify=False)
         return True
 
     def move_position (self, value, relative=True, notify=True):
